@@ -19,14 +19,15 @@
         }
     },
     "require": {
-        "php": "^7.1||^8.0"
+        "php": "^7.4||^8.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~7.0||~9.0"
+        "phpunit/phpunit": "~7.0||~9.0",
+        "phpstan/phpstan": "^0.12.66"
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.x-dev"
+            "dev-master": "3.x-dev"
         }
     }
 }
