@@ -1,8 +1,18 @@
-<?php declare(strict_types=1);
+<?php
+/*
+ * Protocol FCGI library
+ *
+ * @copyright Copyright 2021. Lisachenko Alexander <lisachenko.it@gmail.com>
+ * This source file is subject to the license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+declare(strict_types=1);
 
 namespace Lisachenko\Protocol\FCGI;
 
 use Lisachenko\Protocol\FCGI;
+use ReflectionClass;
 
 /**
  * FCGI record.
@@ -11,73 +21,61 @@ use Lisachenko\Protocol\FCGI;
  */
 class Record
 {
-
     /**
      * Identifies the FastCGI protocol version.
-     *
-     * @var int
      */
-    protected $version = FCGI::VERSION_1;
+    protected int $version = FCGI::VERSION_1;
 
     /**
      * Identifies the FastCGI record type, i.e. the general function that the record performs.
-     *
-     * @var int
      */
-    protected $type = FCGI::UNKNOWN_TYPE;
+    protected int $type = FCGI::UNKNOWN_TYPE;
 
     /**
      * Identifies the FastCGI request to which the record belongs.
-     *
-     * @var int
      */
-    protected $requestId = FCGI::NULL_REQUEST_ID;
+    protected int $requestId = FCGI::NULL_REQUEST_ID;
 
     /**
      * Reserved byte for future proposes
-     *
-     * @var int
      */
-    protected $reserved = 0;
+    protected int $reserved = 0;
 
     /**
      * The number of bytes in the contentData component of the record.
-     *
-     * @var int
      */
-    private $contentLength = 0;
+    private int $contentLength = 0;
 
     /**
      * The number of bytes in the paddingData component of the record.
-     *
-     * @var int
      */
-    private $paddingLength = 0;
+    private int $paddingLength = 0;
 
     /**
      * Binary data, between 0 and 65535 bytes of data, interpreted according to the record type.
-     *
-     * @var string
      */
-    private $contentData = '';
+    private string $contentData = '';
 
     /**
      * Padding data, between 0 and 255 bytes of data, which are ignored.
-     *
-     * @var string
      */
-    private $paddingData = '';
+    private string $paddingData = '';
 
     /**
      * Unpacks the message from the binary data buffer
      *
-     * @param string $data Binary buffer with raw data
-     *
      * @return static
      */
-    final public static function unpack(string $data): self
+    final public static function unpack(string $binaryData): self
     {
-        $self = new static();
+        /** @var static $self */
+        $self   = (new ReflectionClass(static::class))->newInstanceWithoutConstructor();
+
+        /** @phpstan-var false|array{version: int, type: int, requestId: int, contentLength: int, paddingLength: int} */
+        $packet = unpack(FCGI::HEADER_FORMAT, $binaryData);
+        if ($packet === false) {
+            throw new \RuntimeException('Can not unpack data from the binary buffer');
+        }
         [
             $self->version,
             $self->type,
@@ -85,11 +83,11 @@ class Record
             $self->contentLength,
             $self->paddingLength,
             $self->reserved
-        ] = array_values(unpack(FCGI::HEADER_FORMAT, $data));
+        ] = array_values($packet);
 
-        $payload = substr($data, FCGI::HEADER_LEN);
+        $payload = substr($binaryData, FCGI::HEADER_LEN);
         self::unpackPayload($self, $payload);
-        if (get_called_class() !== __CLASS__ && $self->contentLength > 0) {
+        if (static::class !== self::class && $self->contentLength > 0) {
             static::unpackPayload($self, $payload);
         }
 
@@ -119,17 +117,14 @@ class Record
 
     /**
      * Sets the content data and adjusts the length fields
-     *
-     * @param string $data
-     *
-     * @return static
      */
     public function setContentData(string $data): self
     {
-        $this->contentData = $data;
+        $this->contentData   = $data;
         $this->contentLength = strlen($this->contentData);
-        $extraLength = $this->contentLength % 8;
+        $extraLength         = $this->contentLength % 8;
         $this->paddingLength = $extraLength ? (8 - $extraLength) : 0;
+
         return $this;
     }
 
@@ -170,14 +165,11 @@ class Record
      *
      * There should be only one unique ID for all active requests,
      * use random number or preferably resetting auto-increment.
-     *
-     * @param int $requestId
-     *
-     * @return static
      */
     public function setRequestId(int $requestId): self
     {
         $this->requestId = $requestId;
+
         return $this;
     }
 
@@ -201,18 +193,19 @@ class Record
      * Method to unpack the payload for the record.
      *
      * NB: Default implementation will be always called
-     *
-     * @param static $self Instance of current frame
-     * @param string $data Binary data
+     * @param static $self
      */
-    protected static function unpackPayload($self, string $data): void
+    protected static function unpackPayload(Record $self, string $binaryData): void
     {
+        /** @phpstan-var false|array{contentData: string, paddingData: string} */
+        $payload = unpack("a{$self->contentLength}contentData/a{$self->paddingLength}paddingData", $binaryData);
+        if ($payload === false) {
+            throw new \RuntimeException('Can not unpack data from the binary buffer');
+        }
         [
             $self->contentData,
             $self->paddingData
-        ] = array_values(
-            unpack("a{$self->contentLength}contentData/a{$self->paddingLength}paddingData", $data)
-        );
+        ] = array_values($payload);
     }
 
     /**
@@ -222,5 +215,4 @@ class Record
     {
         return pack("a{$this->contentLength}", $this->contentData);
     }
-
 }
