@@ -1,4 +1,13 @@
-<?php declare(strict_types=1);
+<?php
+/*
+ * Protocol FCGI library
+ *
+ * @copyright Copyright 2021. Lisachenko Alexander <lisachenko.it@gmail.com>
+ * This source file is subject to the license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+declare(strict_types=1);
 
 namespace Lisachenko\Protocol\FCGI\Record;
 
@@ -12,17 +21,14 @@ use Lisachenko\Protocol\FCGI\Record;
  */
 class BeginRequest extends Record
 {
-
     /**
      * The role component sets the role the Web server expects the application to play.
      * The currently-defined roles are:
      *   FCGI_RESPONDER
      *   FCGI_AUTHORIZER
      *   FCGI_FILTER
-     *
-     * @var int
      */
-    protected $role = FCGI::UNKNOWN_ROLE;
+    protected int $role = FCGI::UNKNOWN_ROLE;
 
     /**
      * The flags component contains a bit that controls connection shutdown.
@@ -31,23 +37,19 @@ class BeginRequest extends Record
      *   If zero, the application closes the connection after responding to this request.
      *   If not zero, the application does not close the connection after responding to this request;
      *   the Web server retains responsibility for the connection.
-     *
-     * @var int
      */
-    protected $flags;
+    protected int $flags;
 
     /**
      * Reserved data, 5 bytes maximum
-     *
-     * @var string
      */
-    protected $reserved1;
+    protected string $reserved1;
 
     public function __construct(int $role = FCGI::UNKNOWN_ROLE, int $flags = 0, string $reserved = '')
     {
-        $this->type = FCGI::BEGIN_REQUEST;
-        $this->role = $role;
-        $this->flags = $flags;
+        $this->type      = FCGI::BEGIN_REQUEST;
+        $this->role      = $role;
+        $this->flags     = $flags;
         $this->reserved1 = $reserved;
         $this->setContentData($this->packPayload());
     }
@@ -83,18 +85,26 @@ class BeginRequest extends Record
 
     /**
      * {@inheritdoc}
-     * @param static $self
      */
-    protected static function unpackPayload($self, string $data): void
+    protected static function unpackPayload($self, string $binaryData): void
     {
+        assert($self instanceof self);
+
+        /** @phpstan-var false|array{role: int, flags: int, reserved: string} */
+        $payload = unpack("nrole/Cflags/a5reserved", $binaryData);
+        if ($payload === false) {
+            throw new \RuntimeException('Can not unpack data from the binary buffer');
+        }
         [
             $self->role,
             $self->flags,
             $self->reserved1
-        ] = array_values(unpack("nrole/Cflags/a5reserved", $data));
+        ] = array_values($payload);
     }
 
-    /** {@inheritdoc} */
+    /**
+     * {@inheritdoc}
+     */
     protected function packPayload(): string
     {
         return pack(
