@@ -1,4 +1,13 @@
-<?php declare(strict_types=1);
+<?php
+/*
+ * Protocol FCGI library
+ *
+ * @copyright Copyright 2021. Lisachenko Alexander <lisachenko.it@gmail.com>
+ * This source file is subject to the license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+declare(strict_types=1);
 
 namespace Lisachenko\Protocol\FCGI\Record;
 
@@ -13,13 +22,10 @@ use Lisachenko\Protocol\FCGI\Record;
  */
 class EndRequest extends Record
 {
-
     /**
      * The appStatus component is an application-level status code. Each role documents its usage of appStatus.
-     *
-     * @var int
      */
-    protected $appStatus = 0;
+    protected int $appStatus = 0;
 
     /**
      * The protocolStatus component is a protocol-level status code.
@@ -33,24 +39,20 @@ class EndRequest extends Record
      *      This happens when the application runs out of some resource, e.g. database connections.
      *   FCGI_UNKNOWN_ROLE: rejecting a new request.
      *      This happens when the Web server has specified a role that is unknown to the application.
-     *
-     * @var int
      */
-    protected $protocolStatus = FCGI::REQUEST_COMPLETE;
+    protected int $protocolStatus = FCGI::REQUEST_COMPLETE;
 
     /**
      * Reserved data, 3 bytes maximum
-     *
-     * @var string
      */
-    protected $reserved1;
+    protected string $reserved1;
 
     public function __construct(int $protocolStatus = FCGI::REQUEST_COMPLETE, int $appStatus = 0, string $reserved = '')
     {
-        $this->type = FCGI::END_REQUEST;
+        $this->type           = FCGI::END_REQUEST;
         $this->protocolStatus = $protocolStatus;
-        $this->appStatus = $appStatus;
-        $this->reserved1 = $reserved;
+        $this->appStatus      = $appStatus;
+        $this->reserved1      = $reserved;
         $this->setContentData($this->packPayload());
     }
 
@@ -84,18 +86,26 @@ class EndRequest extends Record
 
     /**
      * {@inheritdoc}
-     * @param static $self
      */
-    protected static function unpackPayload($self, string $data): void
+    protected static function unpackPayload($self, string $binaryData): void
     {
+        assert($self instanceof self);
+
+        /** @phpstan-var false|array{appStatus: int, protocolStatus: int, reserved: string} */
+        $payload = unpack("NappStatus/CprotocolStatus/a3reserved", $binaryData);
+        if ($payload === false) {
+            throw new \RuntimeException('Can not unpack data from the binary buffer');
+        }
         [
             $self->appStatus,
             $self->protocolStatus,
             $self->reserved1
-        ] = array_values(unpack("NappStatus/CprotocolStatus/a3reserved", $data));
+        ] = array_values($payload);
     }
 
-    /** {@inheritdoc} */
+    /**
+     * {@inheritdoc}
+     */
     protected function packPayload(): string
     {
         return pack(
@@ -105,5 +115,4 @@ class EndRequest extends Record
             $this->reserved1
         );
     }
-
 }
