@@ -1,4 +1,13 @@
-<?php declare(strict_types=1);
+<?php
+/*
+ * Protocol FCGI library
+ *
+ * @copyright Copyright 2021. Lisachenko Alexander <lisachenko.it@gmail.com>
+ * This source file is subject to the license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+declare(strict_types=1);
 
 namespace Lisachenko\Protocol\FCGI\Record;
 
@@ -17,25 +26,20 @@ use Lisachenko\Protocol\FCGI\Record;
  */
 class UnknownType extends Record
 {
-
     /**
      * Type of the unrecognized management record.
-     *
-     * @var int
      */
-    protected $type1;
+    protected int $type1;
 
     /**
      * Reserved data, 7 bytes maximum
-     *
-     * @var string
      */
-    protected $reserved1;
+    protected string $reserved1;
 
     public function __construct(int $type = 0, string $reserved = '')
     {
-        $this->type = FCGI::UNKNOWN_TYPE;
-        $this->type1 = $type;
+        $this->type      = FCGI::UNKNOWN_TYPE;
+        $this->type1     = $type;
         $this->reserved1 = $reserved;
         $this->setContentData($this->packPayload());
     }
@@ -50,14 +54,22 @@ class UnknownType extends Record
 
     /**
      * {@inheritdoc}
-     * @param static $self
      */
-    public static function unpackPayload($self, string $data): void
+    public static function unpackPayload($self, string $binaryData): void
     {
-        [$self->type1, $self->reserved1] = array_values(unpack("Ctype/a7reserved", $data));
+        assert($self instanceof self);
+
+        /** @phpstan-var false|array{type: int, reserved: string} */
+        $payload = unpack("Ctype/a7reserved", $binaryData);
+        if ($payload === false) {
+            throw new \RuntimeException('Can not unpack data from the binary buffer');
+        }
+        [$self->type1, $self->reserved1] = array_values($payload);
     }
 
-    /** {@inheritdoc} */
+    /**
+     * {@inheritdoc}
+     */
     protected function packPayload(): string
     {
         return pack(
@@ -66,5 +78,4 @@ class UnknownType extends Record
             $this->reserved1
         );
     }
-
 }
