@@ -1,4 +1,13 @@
-<?php declare(strict_types=1);
+<?php
+/*
+ * Protocol FCGI library
+ *
+ * @copyright Copyright 2021. Lisachenko Alexander <lisachenko.it@gmail.com>
+ * This source file is subject to the license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+declare(strict_types=1);
 
 namespace Lisachenko\Protocol\FCGI\Record;
 
@@ -12,7 +21,6 @@ use Lisachenko\Protocol\FCGI\Record;
  */
 class AbortRequest extends Record
 {
-
     public function __construct(int $requestId = 0)
     {
         $this->type = FCGI::ABORT_REQUEST;
