@@ -1,4 +1,13 @@
-<?php declare(strict_types=1);
+<?php
+/*
+ * Protocol FCGI library
+ *
+ * @copyright Copyright 2021. Lisachenko Alexander <lisachenko.it@gmail.com>
+ * This source file is subject to the license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+declare(strict_types=1);
 
 namespace Lisachenko\Protocol\FCGI\Record;
 
@@ -14,11 +23,9 @@ use Lisachenko\Protocol\FCGI\Record;
  */
 class Data extends Record
 {
-
     public function __construct(string $contentData = '')
     {
         $this->type = FCGI::DATA;
         $this->setContentData($contentData);
     }
-
 }
