@@ -1,4 +1,13 @@
-<?php declare(strict_types=1);
+<?php
+/*
+ * Protocol FCGI library
+ *
+ * @copyright Copyright 2021. Lisachenko Alexander <lisachenko.it@gmail.com>
+ * This source file is subject to the license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+declare(strict_types=1);
 
 namespace Lisachenko\Protocol\FCGI\Record;
 
@@ -10,7 +19,7 @@ use Lisachenko\Protocol\FCGI;
  */
 class EndRequestTest extends TestCase
 {
-    protected static $rawMessage = '01030000000800000000006400000000';
+    protected static string $rawMessage = '01030000000800000000006400000000';
 
     public function testPacking(): void
     {
@@ -24,7 +33,9 @@ class EndRequestTest extends TestCase
 
     public function testUnpacking(): void
     {
-        $request = EndRequest::unpack(hex2bin(self::$rawMessage));
+        /** @var string $binaryData */
+        $binaryData = hex2bin(self::$rawMessage);
+        $request    = EndRequest::unpack($binaryData);
 
         $this->assertEquals(FCGI::END_REQUEST, $request->getType());
         $this->assertEquals(FCGI::REQUEST_COMPLETE, $request->getProtocolStatus());
