@@ -1,4 +1,13 @@
-<?php declare(strict_types=1);
+<?php
+/*
+ * Protocol FCGI library
+ *
+ * @copyright Copyright 2021. Lisachenko Alexander <lisachenko.it@gmail.com>
+ * This source file is subject to the license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+declare(strict_types=1);
 
 namespace Lisachenko\Protocol\FCGI;
 
@@ -10,24 +19,24 @@ use Lisachenko\Protocol\FCGI;
  */
 class RecordTest extends TestCase
 {
-
     // from the wireshark captured traffic
-    static $rawRequest = '01010001000800000001010000000000';
+    static string $rawRequest = '01010001000800000001010000000000';
 
     public function testUnpackingPacket(): void
     {
+        /** @var string $packet */
         $packet = hex2bin(self::$rawRequest);
         $record = Record::unpack($packet);
 
         // Verify all general fields
-        $this->assertEquals($record->getVersion(), FCGI::VERSION_1);
-        $this->assertEquals($record->getType(), FCGI::BEGIN_REQUEST);
-        $this->assertEquals($record->getRequestId(), 1);
-        $this->assertEquals($record->getContentLength(), 8);
-        $this->assertEquals($record->getPaddingLength(), 0);
+        $this->assertEquals(FCGI::VERSION_1, $record->getVersion());
+        $this->assertEquals(FCGI::BEGIN_REQUEST, $record->getType());
+        $this->assertEquals(1, $record->getRequestId());
+        $this->assertEquals(8, $record->getContentLength());
+        $this->assertEquals(0, $record->getPaddingLength());
 
         // Check payload data
-        $this->assertEquals($record->getContentData(), hex2bin('0001010000000000'));
+        $this->assertEquals(hex2bin('0001010000000000'), $record->getContentData());
     }
 
     public function testPackingPacket(): void
@@ -39,9 +48,9 @@ class RecordTest extends TestCase
 
         $this->assertEquals($packet, hex2bin('010b0005000503003132333435000000'));
         $result = Record::unpack($packet);
-        $this->assertEquals($result->getType(), FCGI::UNKNOWN_TYPE);
-        $this->assertEquals($result->getRequestId(), 5);
-        $this->assertEquals($result->getContentData(), '12345');
+        $this->assertEquals(FCGI::UNKNOWN_TYPE, $result->getType());
+        $this->assertEquals(5, $result->getRequestId());
+        $this->assertEquals('12345', $result->getContentData());
     }
 
     /**
@@ -51,10 +60,9 @@ class RecordTest extends TestCase
     {
         $record = new Record();
         $record->setContentData('12345');
-        $this->assertEquals($record->getPaddingLength(), 3);
+        $this->assertEquals(3, $record->getPaddingLength());
 
         $record->setContentData('12345678');
-        $this->assertEquals($record->getPaddingLength(), 0);
+        $this->assertEquals(0, $record->getPaddingLength());
     }
-
 }
